requirements.txt
joblib==1.4.2
matplotlib==3.8.0
numpy==1.26.4
pandas==2.2.3
Pillow==11.1.0
plotly==5.17.0
seaborn==0.13.2
streamlit==1.40.2
